Update platform branch with top of tree.
Major fixes:
- Made the "vFile:" GDB remote I/O packets work again and back them with the correct flags for open()
- Improved the security of launching the remote GDB server so it only accepts connections from the host that is connected to the remote platform
- Fixed the platform file commands to use the right arguments so they work on 32/64 bit systems (GetFile()/PutFile()/OpenFile()/CloseFile(), etc)
- Added error objects to the Platform file APIs

llvm-svn: 183468

+++ lldb/include/lldb/Core/StreamBuffer.h
@@ -47,7 +47,7 @@ public:
         // Nothing to do when flushing a buffer based stream...
     }
 
-    virtual int
+    virtual size_t
     Write (const void *s, size_t length)
     {
         if (s && length)
@@ -61,6 +61,10 @@ public:
         m_packet.clear();
     }
 
+    // Beware, this might not be NULL terminated as you can expect from
+    // StringString as there may be random bits in the llvm::SmallVector. If
+    // you are using this class to create a C string, be sure the call PutChar ('\0')
+    // after you have created your string, or use StreamString.
     const char *
     GetData () const
     {
